favor uint256 over uint

+++ contracts/OneTokenFactory.sol
@@ -272,14 +272,14 @@ contract OneTokenFactory is IOneTokenFactory, ICHICommon {
     /**
      @notice returns the count of deployed and initialized oneToken instances
      */
-    function oneTokenCount() external view override returns(uint) {
+    function oneTokenCount() external view override returns(uint256) {
         return oneTokenSet.count();
     }
 
     /**
      @notice returns the address of the deployed/initialized oneToken instance at the index
      */
-    function oneTokenAtIndex(uint index) external view override returns(address) {
+    function oneTokenAtIndex(uint256 index) external view override returns(address) {
         return oneTokenSet.keyAtIndex(index);
     }
 
@@ -295,14 +295,14 @@ contract OneTokenFactory is IOneTokenFactory, ICHICommon {
     /**
      @notice returns the count of the registered modules
      */
-    function moduleCount() external view override returns(uint) {
+    function moduleCount() external view override returns(uint256) {
         return moduleSet.count();
     }
 
     /**
      @notice returns the address of the registered module at the index
      */
-    function moduleAtIndex(uint index) external view override returns(address module) {
+    function moduleAtIndex(uint256 index) external view override returns(address module) {
         return moduleSet.keyAtIndex(index);
     }
 
@@ -357,21 +357,21 @@ contract OneTokenFactory is IOneTokenFactory, ICHICommon {
      @notice returns count of foreignTokens registered with the factory
      @dev includes memberTokens, otherTokens and collateral tokens but not oneTokens
      */
-    function foreignTokenCount() external view override returns(uint) {
+    function foreignTokenCount() external view override returns(uint256) {
         return foreignTokenSet.count();
     }
 
     /**
      @notice returns the address of the foreignToken at the index
      */
-    function foreignTokenAtIndex(uint index) external view override returns(address) {
+    function foreignTokenAtIndex(uint256 index) external view override returns(address) {
         return foreignTokenSet.keyAtIndex(index);
     }
 
     /**
      @notice returns foreignToken metadata for the given foreignToken
      */
-    function foreignTokenInfo(address foreignToken) external view override returns(bool collateral, uint oracleCount) {
+    function foreignTokenInfo(address foreignToken) external view override returns(bool collateral, uint256 oracleCount) {
         ForeignToken storage f = foreignTokens[foreignToken];
         collateral = f.isCollateral;
         oracleCount = f.oracleSet.count();
@@ -380,14 +380,14 @@ contract OneTokenFactory is IOneTokenFactory, ICHICommon {
     /**
      @notice returns the count of oracles registered for the given foreignToken
      */
-    function foreignTokenOracleCount(address foreignToken) external view override returns(uint) {
+    function foreignTokenOracleCount(address foreignToken) external view override returns(uint256) {
         return foreignTokens[foreignToken].oracleSet.count();
     }
 
     /**
      @notice returns the foreignToken oracle address at the index
      */
-    function foreignTokenOracleAtIndex(address foreignToken, uint index) external view override returns(address) {
+    function foreignTokenOracleAtIndex(address foreignToken, uint256 index) external view override returns(address) {
         return foreignTokens[foreignToken].oracleSet.keyAtIndex(index);
     }
