+++ contracts/_uniswap/v2-core/contracts/UniswapV2Factory.sol
@@ -12,13 +12,13 @@ contract UniswapV2Factory is IUniswapV2Factory {
     mapping(address => mapping(address => address)) public getPair;
     address[] public allPairs;
 
-    event PairCreated(address indexed token0, address indexed token1, address pair, uint);
+    event PairCreated(address indexed token0, address indexed token1, address pair, uint256);
 
     constructor(address _feeToSetter) {
         feeToSetter = _feeToSetter;
     }
 
-    function allPairsLength() override external view returns (uint) {
+    function allPairsLength() override external view returns (uint256) {
         return allPairs.length;
     }
