+++ contracts/interface/IOneTokenFactory.sol
@@ -39,25 +39,25 @@ interface IOneTokenFactory is InterfaceCommon {
     
     function MODULE_TYPE() external view returns(bytes32);
 
-    function oneTokenCount() external view returns(uint);
-    function oneTokenAtIndex(uint index) external view returns(address);
+    function oneTokenCount() external view returns(uint256);
+    function oneTokenAtIndex(uint256 index) external view returns(address);
     function isOneToken(address oneToken) external view returns(bool);
  
     // modules
 
-    function moduleCount() external view returns(uint);
-    function moduleAtIndex(uint index) external view returns(address module);
+    function moduleCount() external view returns(uint256);
+    function moduleAtIndex(uint256 index) external view returns(address module);
     function moduleInfo(address module) external view returns(string memory name, string memory url, ModuleType moduleType);
     function isModule(address module) external view returns(bool);
     function isValidModuleType(address module, ModuleType moduleType) external view returns(bool);
 
     // foreign tokens
 
-    function foreignTokenCount() external view returns(uint);
-    function foreignTokenAtIndex(uint index) external view returns(address);
-    function foreignTokenInfo(address foreignToken) external view returns(bool collateral, uint oracleCount);
-    function foreignTokenOracleCount(address foreignToken) external view returns(uint);
-    function foreignTokenOracleAtIndex(address foreignToken, uint index) external view returns(address);
+    function foreignTokenCount() external view returns(uint256);
+    function foreignTokenAtIndex(uint256 index) external view returns(address);
+    function foreignTokenInfo(address foreignToken) external view returns(bool collateral, uint256 oracleCount);
+    function foreignTokenOracleCount(address foreignToken) external view returns(uint256);
+    function foreignTokenOracleAtIndex(address foreignToken, uint256 index) external view returns(address);
     function isOracle(address foreignToken, address oracle) external view returns(bool);
     function isForeignToken(address foreignToken) external view returns(bool);
     function isCollateral(address foreignToken) external view returns(bool);
