+++ contracts/interface/IOneTokenV1.sol
@@ -7,16 +7,16 @@ import "./IOneTokenV1Base.sol";
 
 interface IOneTokenV1 is IOneTokenV1Base {
 
-    function mintingFee() external view returns(uint);
-    function redemptionFee() external view returns(uint);
-    // function withdraw(address token, uint amount) external;
-    function mint(address collateral, uint oneTokens) external;
-    function redeem(address collateral, uint amount) external;
-    function setMintingFee(uint fee) external;
-    function setRedemptionFee(uint fee) external;
-    function updateMintingRatio(address collateralToken) external returns(uint ratio, uint maxOrderVolume);
-    // function userBalances(address, address) external view returns(uint);
-    // function userCreditBlocks(address, address) external view returns(uint);
-    function getMintingRatio(address collateralToken) external view returns(uint ratio, uint maxOrderVolume);
-    function getHoldings(address token) external view returns(uint vaultBalance, uint strategyBalance);
+    function mintingFee() external view returns(uint256);
+    function redemptionFee() external view returns(uint256);
+    // function withdraw(address token, uint256 amount) external;
+    function mint(address collateral, uint256 oneTokens) external;
+    function redeem(address collateral, uint256 amount) external;
+    function setMintingFee(uint256 fee) external;
+    function setRedemptionFee(uint256 fee) external;
+    function updateMintingRatio(address collateralToken) external returns(uint256 ratio, uint256 maxOrderVolume);
+    // function userBalances(address, address) external view returns(uint256);
+    // function userCreditBlocks(address, address) external view returns(uint256);
+    function getMintingRatio(address collateralToken) external view returns(uint256 ratio, uint256 maxOrderVolume);
+    function getHoldings(address token) external view returns(uint256 vaultBalance, uint256 strategyBalance);
 }
